requirements.txt.txt
line-bot-sdk
flask
gunicorn
python-dotenv
lxml
requests
pandas
